// src/types/game.ts
export type Suit = number /* int */
export const Hearts: Suit = 0
export const Diamonds: Suit = 1
export const Clubs: Suit = 2
export const Spades: Suit = 3
export type Rank = number /* int */
export const Four = 0
export const Five = 1
export const Six = 2
export const Seven = 3
export const Eight = 4
export const Nine = 5
export const Ten = 6
export const Jack = 7
export const Queen = 8
export const King = 9
export const Ace = 10
export const Two = 11
export const Joker = 12
export const Three = 13
export const Wild = 14
export interface Card {
  id: number /* int */
  suit: Suit
  rank: Rank
}